+++ packages/apollo-storybook-react/example/.storybook/config.js
@@ -3,10 +3,35 @@ import apolloStorybookDecorator from '../../src';
 import typeDefs from '../schema/typeDefinitions';
 import mocks from '../schema/mocks';
 
+import { withClientState } from 'apollo-link-state';
+
 addDecorator(
   apolloStorybookDecorator({
     typeDefs,
     mocks,
+    links: (cache) => {
+      const linkState = withClientState({
+        cache,
+        resolvers: {
+          Mutation: {
+            updateNetworkStatus: (_, { isConnected }, { cache }) => {
+              const data = {
+                networkStatus: {
+                  __typename: 'NetworkStatus',
+                  isConnected
+                },
+              };
+              cache.writeData({ data });
+              return null
+            },
+          },
+        }
+      });
+
+      return [
+        linkState,
+      ];
+    }
   })
 );