+++ packages/apollo-storybook-react/src/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import { ApolloProvider } from 'react-apollo';
 import createClient from 'apollo-storybook-core';
 
@@ -7,12 +7,14 @@ export default function initializeApollo({
   mocks,
   apolloLinkOptions = {},
   apolloClientOptions = {},
+  links,
   typeResolvers,
   context = {},
   rootValue = {},
   // cacheOptions is a necessary config parameter because some use cases will require a pre-configured
   // fragmentMatcher such as IntrospectionFragmentMatcher, etc.
   cacheOptions = {},
+  resolverValidationOptions,
 }) {
   const graphqlClient = createClient({
     mocks,
@@ -22,13 +24,15 @@ export default function initializeApollo({
     typeDefs,
     rootValue,
     context,
+    links,
     cacheOptions,
+    resolverValidationOptions,
   });
 
   function StorybookProvider({ children }) {
     return (
       <ApolloProvider client={graphqlClient}>
-        <div>{children}</div>
+        <Fragment>{children}</Fragment>
       </ApolloProvider>
     );
   }