+++ packages/apollo-storybook-core/CHANGELOG.md
@@ -3,6 +3,124 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.2"></a>
+## [0.5.2](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.5.1...apollo-storybook-core@0.5.2) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-core
+
+<a name="0.5.1"></a>
+## [0.5.1](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.3.4...apollo-storybook-core@0.5.1) (2018-08-20)
+
+
+### Bug Fixes
+
+* **remove:** stage0 ([a1b7d20](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/a1b7d20))
+
+
+
+
+<a name="0.5.0"></a>
+# [0.5.0](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/v0.5.1...v0.5.0) (2018-08-20)
+
+
+### Bug Fixes
+
+* remove wrong and redundant links default value ([2f78f52](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/2f78f52))
+* **peer:** peer deps ([fa43be8](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/fa43be8))
+* **preset:** preset env ([6a45b6f](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/6a45b6f))
+* **preset:** preset env ([c85631d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c85631d))
+* **remove:** stage0 ([a1b7d20](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/a1b7d20))
+* **resolve:** update core library ([b07cd92](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/b07cd92))
+* **storybook:** storybook links ([e121164](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/e121164))
+
+
+### Features
+
+* **apollo:** 2.0 ([c01c69a](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c01c69a))
+* **apollo:** 2.0 ([578426d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/578426d))
+* **lerna:** add lerna to the project ([5fc9345](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/5fc9345))
+* **lerna:** add lerna to the project ([84e6f4d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/84e6f4d))
+* **Links:** pass links through to storybook core ([8b97087](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/8b97087))
+
+
+
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/v0.5.1...v0.4.0) (2018-08-20)
+
+
+### Bug Fixes
+
+* remove wrong and redundant links default value ([2f78f52](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/2f78f52))
+* **peer:** peer deps ([fa43be8](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/fa43be8))
+* **preset:** preset env ([6a45b6f](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/6a45b6f))
+* **preset:** preset env ([c85631d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c85631d))
+* **remove:** stage0 ([a1b7d20](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/a1b7d20))
+* **resolve:** update core library ([b07cd92](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/b07cd92))
+* **storybook:** storybook links ([e121164](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/e121164))
+
+
+### Features
+
+* **apollo:** 2.0 ([c01c69a](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c01c69a))
+* **apollo:** 2.0 ([578426d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/578426d))
+* **lerna:** add lerna to the project ([5fc9345](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/5fc9345))
+* **lerna:** add lerna to the project ([84e6f4d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/84e6f4d))
+* **Links:** pass links through to storybook core ([8b97087](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/8b97087))
+
+
+
+<a name="0.3.4"></a>
+## [0.3.4](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.3.3...apollo-storybook-core@0.3.4) (2018-06-09)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-core
+
+<a name="0.3.3"></a>
+## [0.3.3](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.3.2...apollo-storybook-core@0.3.3) (2018-05-15)
+
+
+### Bug Fixes
+
+* **preset:** preset env ([c85631d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c85631d))
+
+
+
+
+<a name="0.3.2"></a>
+## [0.3.2](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.3.1...apollo-storybook-core@0.3.2) (2018-04-25)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-core
+
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.3.0...apollo-storybook-core@0.3.1) (2018-04-25)
+
+
+### Bug Fixes
+
+* **storybook:** storybook links ([e121164](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/e121164))
+
+
+
+
+<a name="0.3.0"></a>
+# 0.3.0 (2018-04-25)
+
+
+### Features
+
+* **Links:** pass links through to storybook core ([8b97087](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/8b97087))
+
+
+
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-core@0.1.1...apollo-storybook-core@0.2.0) (2017-12-26)