+++ packages/apollo-storybook-react-native/CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.8"></a>
+## [0.2.8](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.7...apollo-storybook-react-native@0.2.8) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react-native
+
+<a name="0.2.7"></a>
+## [0.2.7](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.6...apollo-storybook-react-native@0.2.7) (2018-08-20)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react-native
+
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.5...apollo-storybook-react-native@0.2.6) (2018-08-20)
+
+
+### Bug Fixes
+
+* **remove:** stage0 ([16914c1](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/16914c1))
+
+
+
+
+<a name="0.2.5"></a>
+## [0.2.5](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.4...apollo-storybook-react-native@0.2.5) (2018-06-09)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react-native
+
+<a name="0.2.4"></a>
+## [0.2.4](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.3...apollo-storybook-react-native@0.2.4) (2018-05-30)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react-native
+
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.2...apollo-storybook-react-native@0.2.3) (2018-05-15)
+
+
+### Bug Fixes
+
+* **resolve:** update core library ([b07cd92](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/b07cd92))
+
+
+
+
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react-native@0.2.1...apollo-storybook-react-native@0.2.2) (2018-05-15)
+
+
+### Bug Fixes
+
+* **preset:** preset env ([c85631d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c85631d))
+
+
+
+
 <a name="0.2.1"></a>
 ## 0.2.1 (2017-12-07)