+++ packages/apollo-storybook-react/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.8"></a>
+## [0.1.8](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.7...apollo-storybook-react@0.1.8) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react
+
+<a name="0.1.7"></a>
+## [0.1.7](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.6...apollo-storybook-react@0.1.7) (2018-08-20)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react
+
+<a name="0.1.6"></a>
+## [0.1.6](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.5...apollo-storybook-react@0.1.6) (2018-08-20)
+
+
+### Bug Fixes
+
+* **remove:** stage0 ([16914c1](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/16914c1))
+
+
+
+
+<a name="0.1.5"></a>
+## [0.1.5](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.4...apollo-storybook-react@0.1.5) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package apollo-storybook-react
+
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.3...apollo-storybook-react@0.1.4) (2018-06-09)
+
+
+### Bug Fixes
+
+* **peer:** peer deps ([fa43be8](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/fa43be8))
+
+
+
+
+<a name="0.1.3"></a>
+## [0.1.3](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.2...apollo-storybook-react@0.1.3) (2018-05-30)
+
+
+### Bug Fixes
+
+* remove wrong and redundant links default value ([2f78f52](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/2f78f52))
+
+
+
+
+<a name="0.1.2"></a>
+## [0.1.2](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.1...apollo-storybook-react@0.1.2) (2018-05-15)
+
+
+### Bug Fixes
+
+* **resolve:** update core library ([b07cd92](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/b07cd92))
+
+
+
+
+<a name="0.1.1"></a>
+## [0.1.1](https://github.com/abhiaiyer91/apollo-storybook-decorator/compare/apollo-storybook-react@0.1.0...apollo-storybook-react@0.1.1) (2018-05-15)
+
+
+### Bug Fixes
+
+* **preset:** preset env ([c85631d](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/c85631d))
+* **storybook:** storybook links ([e121164](https://github.com/abhiaiyer91/apollo-storybook-decorator/commit/e121164))
+
+
+
+
 <a name="0.1.0"></a>
 # 0.1.0 (2017-11-19)